optimisations
Completely untested!  Could be broken. Just wanted to get this stage captured for posterity.
There seemed to be a bug in the seek, which I've fixed. No idea how it managed to work before - but *actually* I realised this may actually explain the half-frame issue I'vebeen seeing after seeking. Let's see if that is gone.